feat: Refactor graph query handler and degen airdrop network scripts

+++ graph_query/query_degen_airdrop_network.py
@@ -2,6 +2,7 @@
 import pandas as pd
 import random
 import json
+import os
 
 from graph_query.graph_query_handler import GraphQueryHandler
 from utils.costum_keys import CustomKeys as ck
@@ -19,10 +20,12 @@ def save_addresses_with_stats(
         handler: GraphQueryHandler, 
         name: str = 'airdrop1',
     ):
-    stats = handler.query_wallet_stats_as_df(addresses)
     path = f'{DEGEN_AIRDROP_FOLDER}/network/{name}_with_stats.csv'
+    if os.path.exists(path):
+        return pd.read_csv(path)
+    stats = handler.query_wallet_stats_as_df(addresses)
     stats.to_csv(path, index=False)
-
+    return stats
 
 def get_n_claimers(
         handler: GraphQueryHandler, 
@@ -40,24 +43,26 @@ def get_n_claimers(
     else:
         return pd.read_csv(f'{DEGEN_AIRDROP_FOLDER}/network/{claimers_name}_with_stats.csv')
 
+def read_degree_transactions(degree):
+    received_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_received.json'
+    received = json.load(open(received_path, 'r'))
+    sent_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_sent.json'
+    sent = json.load(open(sent_path, 'r'))
+    return received, sent
+
 def save_degree_transactions(
         handler: GraphQueryHandler, 
         addresses: list, 
         contracts: list, 
         degree: str,
     ):
+    received_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_received.json'
+    sent_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_sent.json'
+    if os.path.exists(received_path) and os.path.exists(sent_path):
+        return read_degree_transactions(degree)
     received, sent = handler.get_wallet_transactions(addresses, contracts)
-    received_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_received.json'
     json.dump(received, open(received_path, 'w'))
-    sent_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_sent.json'
     json.dump(sent, open(sent_path, 'w'))
-    return received, sent
-
-def read_degree_transactions(degree):
-    received_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_received.json'
-    received = json.load(open(received_path, 'r'))
-    sent_path = f'{DEGEN_AIRDROP_FOLDER}/network/{degree}_sent.json'
-    sent = json.load(open(sent_path, 'r'))
     return received, sent
 
 def process_degree(
@@ -73,11 +78,11 @@ def process_degree(
     else:
         tnxs = read_degree_transactions(degree)
     received, sent = tnxs
-    senders = [t[ck.FROM_ADDRESS] for tnxs in received.values() for t in tnxs]
-    receivers = [t[ck.TO_ADDRESS] for tnxs in sent.values() for t in tnxs]
-    t = f'{len(senders)} receivers and {len(receivers)} senders for {degree}.'
+    parents = [t[ck.FROM_ADDRESS] for tnxs in received.values() for t in tnxs]
+    children = [t[ck.TO_ADDRESS] for tnxs in sent.values() for t in tnxs]
+    t = f'{len(parents)} parents and {len(children)} children for {degree}.'
     print(t)
-    return senders, receivers
+    return parents, children
 
 def process_layers(
         handler: GraphQueryHandler,
@@ -93,17 +98,17 @@ def process_layers(
     if degree != 0:
             degree_name += '_parent' if parent else '_child'
     degree_args = (handler, source_addresses, contracts, degree_name, True)
-    senders, receivers = process_degree(*degree_args)
-    random.shuffle(senders); random.shuffle(receivers)
-    senders, receivers = senders[:edge_limit], receivers[:edge_limit]
+    parents, children = process_degree(*degree_args)
+    random.shuffle(parents); random.shuffle(children)
+    parents, children = parents[:edge_limit], children[:edge_limit]
     if degree >= degree_limit:
          return
     if parent:
-        save_addresses_with_stats(senders, handler, degree_name)
-        process_layers(handler, senders, contracts, degree + 1, parent)
+        save_addresses_with_stats(parents, handler, degree_name)
+        process_layers(handler, parents, contracts, degree + 1, parent)
     else:
-        save_addresses_with_stats(receivers, handler, degree_name)
-        process_layers(handler, receivers, contracts, degree + 1, parent)
+        save_addresses_with_stats(children, handler, degree_name)
+        process_layers(handler, children, contracts, degree + 1, parent)
     return
 
 def main():
@@ -114,8 +119,9 @@ def main():
     degree_0_addresses = claimers_n['wallet_address'].values
     contracts = ['0x4ed4e862860bed51a9570b96d89af5e1b0efefed']
     # save_addresses_with_stats(degree_0_addresses, handler, 'degree_0')
-    # process_layers(handler, degree_0_addresses, contracts, 0, True, 5, 10)
+    process_layers(handler, degree_0_addresses, contracts, 0, True, 5, 10)
     process_layers(handler, degree_0_addresses, contracts, 0, False, 5, 10)
+    print('Done')
 
 if __name__ == '__main__':
     main()
